All Visual Studio 2010 solution (32-bit only for now). Fix VS2010 compilation warning from Geoid.cpp. Documentation fixes for Albers and LCC. Documentation fixes for Matlab mgrs routines.

+++ src/Geoid.cpp
@@ -234,7 +234,8 @@ namespace GeographicLib {
         if (!(is >> commentid >> key) || commentid != "#")
           continue;
         if (key == "Description" || key =="DateTime") {
-          string::size_type p = s.find_first_not_of(" \t", is.tellg());
+          string::size_type p =
+            s.find_first_not_of(" \t", unsigned(is.tellg()));
           if (p != string::npos)
             (key == "Description" ? _description : _datetime) = s.substr(p);
         } else if (key == "Offset") {
